Source/requirements.txt
streamlit
googlesearch
pandas
numpy
matplotlib
seaborn
yagmail
google-generativeai
pymongo
opencv-python
tensorflow
keras
scikit-learn
textblob
PIL
pyautogui
pathlib
ctypes
comtypes
pycaw
geopy
pyttsx3